break out asyncroute into standalone component
